TODO DOC xetex-synctex: add closing tags
This is an extention to SyncTeX format.
A record /n indicates that input file n has been closed.
No new contents will be produced from this file (though some might be
buffered and not yet shipped out).

Usinc synctex(5) notation, the grammar is extended with a rule:

    <Input Line> ::= "/" <tag> <EOL>

Knowing when a file is closed prevent TeXpresso for needlessly
re-rendering and scanning the remainder of the document.