TODO DOC xetex-pic: hooks for caching bounds of a picture
When a picture is included, XeTeX scans the source file to determine the
bounds of the picture. Picture scanning can become a bottleneck when
used interactively with TeXpresso.

This patch prepares hooks for caching the bounds of a picture. When
implemented, XeTeX saves the bounds on first pass and access them
directly on later passes.