TODO DOC Add a bundle serve command
`tectonic -X bundle serve` is an alternative to `textonic -X bundle cat`
that spawns a long-living process to answer many requests.

It reads newline-delimited filenames on the standard input and outputs
a record for each line a simple binary protocol.

The first character can be 'P' for path, 'C' for contents or 'E' for
error. It is followed the size of the data, serialized as an unsigned
64-bit integer in little-endian, and the data.